+++ src/form/shared.jl
@@ -10,6 +10,32 @@
 # only to prevent code replication
 #
 
+
+""
+function variable_shunt_factor(pm::AbstractWConvexModels; nw::Int=pm.cnw, cnd::Int=pm.ccnd, relax::Bool=false)
+    if relax == true
+        var(pm, nw)[:z_shunt] = JuMP.@variable(pm.model,
+            [i in ids(pm, nw, :shunt)], base_name="$(nw)_z_shunt", 
+            upper_bound = 1, 
+            lower_bound = 0,
+            start = comp_start_value(ref(pm, nw, :shunt, i), "z_shunt_start", 1, 1.0)
+        )
+    else
+        var(pm, nw)[:z_shunt] = JuMP.@variable(pm.model,
+            [i in ids(pm, nw, :shunt)], base_name="$(nw)_z_shunt", 
+            binary = true,
+            start = comp_start_value(ref(pm, nw, :shunt, i), "z_shunt_start", 1, 1.0)
+        )
+    end
+    var(pm, nw, cnd)[:wz_shunt] = JuMP.@variable(pm.model,
+        [i in ids(pm, nw, :shunt)], base_name="$(nw)_wz_shunt",
+        lower_bound = 0,
+        upper_bound = ref(pm, nw, :bus)[ref(pm, nw, :shunt, i)["shunt_bus"]]["vmax"]^2,
+        start = comp_start_value(ref(pm, nw, :shunt, i), "wz_shunt_start", cnd, 1.001)
+    )
+end
+
+
 "`t[ref_bus] == 0`"
 function constraint_theta_ref(pm::AbstractPolarModels, n::Int, c::Int, i::Int)
     JuMP.@constraint(pm.model, var(pm, n, c, :va)[i] == 0)
@@ -84,6 +110,51 @@ function constraint_power_balance(pm::AbstractWModels, n::Int, c::Int, i, bus_ar
         - sum(qd for qd in values(bus_qd))
         + sum(bs for bs in values(bus_bs))*w
     )
+end
+
+
+""
+function constraint_power_balance_ls(pm::AbstractWConvexModels, n::Int, c::Int, i::Int, bus_arcs, bus_arcs_dc, bus_arcs_sw, bus_gens, bus_storage, bus_pd, bus_qd, bus_gs, bus_bs)
+    w    = var(pm, n, c, :w, i)
+    p    = get(var(pm, n, c),    :p, Dict()); _check_var_keys(p, bus_arcs, "active power", "branch")
+    q    = get(var(pm, n, c),    :q, Dict()); _check_var_keys(q, bus_arcs, "reactive power", "branch")
+    pg   = get(var(pm, n, c),   :pg, Dict()); _check_var_keys(pg, bus_gens, "active power", "generator")
+    qg   = get(var(pm, n, c),   :qg, Dict()); _check_var_keys(qg, bus_gens, "reactive power", "generator")
+    ps   = get(var(pm, n, c),   :ps, Dict()); _check_var_keys(ps, bus_storage, "active power", "storage")
+    qs   = get(var(pm, n, c),   :qs, Dict()); _check_var_keys(qs, bus_storage, "reactive power", "storage")
+    psw  = get(var(pm, n, c),  :psw, Dict()); _check_var_keys(psw, bus_arcs_sw, "active power", "switch")
+    qsw  = get(var(pm, n, c),  :qsw, Dict()); _check_var_keys(qsw, bus_arcs_sw, "reactive power", "switch")
+    p_dc = get(var(pm, n, c), :p_dc, Dict()); _check_var_keys(p_dc, bus_arcs_dc, "active power", "dcline")
+    q_dc = get(var(pm, n, c), :q_dc, Dict()); _check_var_keys(q_dc, bus_arcs_dc, "reactive power", "dcline")
+
+    z_demand = get(var(pm, n), :z_demand, Dict()); _check_var_keys(z_demand, keys(bus_pd), "power factor", "load")
+    z_shunt = get(var(pm, n), :z_shunt, Dict()); _check_var_keys(z_shunt, keys(bus_gs), "power factor", "shunt")
+    wz_shunt = get(var(pm, n, c), :wz_shunt, Dict()); _check_var_keys(wz_shunt, keys(bus_gs), "voltage square power factor", "shunt")
+
+    con(pm, n, c, :kcl_p)[i] = JuMP.@constraint(pm.model,
+        sum(p[a] for a in bus_arcs)
+        + sum(p_dc[a_dc] for a_dc in bus_arcs_dc)
+        + sum(psw[a_sw] for a_sw in bus_arcs_sw)
+        ==
+        sum(pg[g] for g in bus_gens)
+        - sum(ps[s] for s in bus_storage)
+        - sum(pd*z_demand[i] for (i,pd) in bus_pd)
+        - sum(gs*wz_shunt[i] for (i,gs) in bus_gs)
+    )
+    con(pm, n, c, :kcl_q)[i] = JuMP.@constraint(pm.model,
+        sum(q[a] for a in bus_arcs)
+        + sum(q_dc[a_dc] for a_dc in bus_arcs_dc)
+        + sum(qsw[a_sw] for a_sw in bus_arcs_sw)
+        ==
+        sum(qg[g] for g in bus_gens)
+        - sum(qs[s] for s in bus_storage)
+        - sum(qd*z_demand[i] for (i,qd) in bus_qd)
+        + sum(bs*wz_shunt[i] for (i,bs) in bus_bs)
+    )
+
+    for s in keys(bus_gs)
+        InfrastructureModels.relaxation_product(pm.model, w, z_shunt[s], wz_shunt[s])
+    end
 end